refactored out stat updater logic
